chore: 🔧 Move RoleType enum in enum package
