+++ src/main/java/com/springboot/api/dto/user/AddUserReq.java
@@ -1,7 +1,7 @@
 package com.springboot.api.dto.user;
 
 
-import com.springboot.api.domain.RoleType;
+import com.springboot.enums.RoleType;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
